pyproject.toml
[tool.poetry]
name = "oneclick-c4d"
version = "0.1.0"
description = ""
authors = ["izzy lyseggen <izzy.lyseggen@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9, <3.10"
specklepy = "^2.6.7"
numpy-stl = "^2.16.3"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pylint = "^2.13.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
